refactor axios to use defaults
