--- client/src/utils/setAuthToken.js
@@ -1,9 +1,0 @@
-import axios from "axios";
-const setAuthToken = token => {
-  if (token) {
-    axios.defaults.headers.common["x-auth-token"] = token;
-  } else {
-    delete axios.defaults.headers.common["x-auth-token"];
-  }
-};
-export default setAuthToken;
